+++ projects/angular-base/src/public_api.ts
@@ -1,5 +1,0 @@
-/*
- * Public API Surface of angular-base
- */
-
-export * from './lib/angular-base.module';