+++ projects/angular-commons/src/public_api.ts
@@ -1,0 +1,5 @@
+/*
+ * Public API Surface of angular-commons
+ */
+
+export * from './lib/angular-commons.module';